requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
Django==1.7
django-appconf==1.0.1
django-compressor==1.5
django-easy-timezones==0.2.0
django-parsley==0.5
django-postgrespool==0.3.0
djangotoolbox==1.6.2
gunicorn==19.2.1
ipython==3.1.0
mysqlclient==1.3.6
nodeenv==0.13.1
oauthlib==0.7.2
psycopg2==2.6
pygeoip==0.3.2
PyJWT==1.1.0
PyMySQL==0.6.6
python-social-auth==0.2.7
python3-openid==3.0.5
pytz==2015.2
requests==2.6.0
requests-oauthlib==0.4.2
six==1.9.0
SQLAlchemy==0.9.8
whitenoise==1.0.6
